refactor JSSmart and use interface
