+++ js-driver/src/main/java/com/jdiai/jsbuilder/JSFilterTemplates.java
@@ -1,6 +1,6 @@
 package com.jdiai.jsbuilder;
 
-public class JSSmartTemplates {
+public class JSFilterTemplates {
     public static String ELEMENT_CONDITION =
         "condition = function (el) {\n" +
         "  if (!el) return false;\n" +
@@ -13,22 +13,27 @@ public class JSSmartTemplates {
         "  if (r.width < 5 || r.height < 5) return false;\n" +
         "  return true;\n" +
         "}\n";
+
     public static String GET_ELEMENT =
         "let start = Date.now();\n" +
         "while (Date.now() - start < 10000) {\n" +
-        "    if (isVisible(document.getElementById('suspend-button'))) break;\n" +
+        "    if (isVisible(document.getElementById('element-id'))) break;\n" +
         "}";
+
     public static String FILTER_FUNC =
         "filter = function(element) {\n" +
         "  return element && getComputedStyle(element).visibility === 'visible';\n" +
         "}\n";
+
     public static String ONE_TO_RESULT = "return %s;\n";
+
     public static String LIST_TO_RESULT =
         "let result = [];\n" +
         "for(let element of elements) {\n" +
         "  result.push(%s);\n" +
         "}\n" +
         "return result;";
+
     public static String ONE_TO_LIST =
         "%s;\n" +
         "list = [];\n" +
@@ -38,7 +43,9 @@ public class JSSmartTemplates {
         "  }\n" +
         "}\n" +
         "elements = Array.from(list.length === 0 ? elements[0] : list);\n";
+
     public static String ONE_TO_ONE = "%s;\n";
+
     public static String LIST_TO_ONE =
         "found = false;\ni = 0;\nfirst = null;\n" +
         "while (!found && i < elements.length) {\n" +
@@ -48,6 +55,7 @@ public class JSSmartTemplates {
         "  i++;\n" +
         "}\n" +
         "if (!found && element) { element = first; }\n";
+
     public static String LIST_TO_LIST =
         "list = [];\nfirst = null;\n" +
         "for(let element of elements) {\n" +
